Refactor kafka and assert package name.
