+++ bitsail-test/bitsail-test-integration/bitsail-test-integration-kafka/src/test/java/com/bytedance/bitsail/test/integration/kafka/FakeToKafkaITCase.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.bytedance.bitsail.test.integration;
+package com.bytedance.bitsail.test.integration.kafka;
 
 import com.bytedance.bitsail.common.configuration.BitSailConfiguration;
 import com.bytedance.bitsail.connector.kafka.option.KafkaWriterOptions;
-import com.bytedance.bitsail.test.integration.kafka.KafkaCluster;
+import com.bytedance.bitsail.test.integration.AbstractIntegrationTest;
 import com.bytedance.bitsail.test.integration.utils.JobConfUtils;
 
 import org.apache.kafka.clients.consumer.ConsumerRecord;
